keylocker/requirements.txt
# Original requirements
pickledb==0.9.2
cryptography==2.4.2
fire==0.1.3
# Added for YAML support
PyYAML>=5.0
